requirements/run.txt
Django<1.10
django-filter
djangorestframework
psycopg2
